+++ internal/config/config.go
@@ -11,6 +11,7 @@ type Config struct {
 	SettingsFileName string       `mapstructure:"settings"`
 	ReposDirectory   string       `mapstructure:"reposDirectory"`
 	RunMode          RunMode      `mapstructure:"runMode"`
+	MaxWorkers       int          `mapstructure:"maxWorkers"`
 	Output           OutputFormat `mapstructure:"output"`
 }